Abstract parser methods into FinanceParserUtil
