+++ src/main/java/seedu/jarvis/logic/parser/finance/EditInstallmentCommandParser.java
@@ -42,14 +42,14 @@ public class EditInstallmentCommandParser implements Parser<EditInstallmentComma
                 new EditInstallmentCommand.EditInstallmentDescriptor();
         if (argMultimap.getValue(PREFIX_DESCRIPTION).isPresent()) {
             editInstallmentDescriptor
-                    .setDescription(ParserUtil
+                    .setDescription(FinanceParserUtil
                             .parseInstallmentDescription(argMultimap
                                     .getValue(PREFIX_DESCRIPTION)
                                     .get()));
         }
         if (argMultimap.getValue(PREFIX_MONEY).isPresent()) {
             editInstallmentDescriptor
-                    .setMoneyPaid(ParserUtil
+                    .setMoneyPaid(FinanceParserUtil
                             .parseInstallmentMoneySpent(argMultimap
                                     .getValue(PREFIX_MONEY)
                                     .get()));
